+++ backend/src/repository/xtream_repository.rs
@@ -1,8 +1,8 @@
-use shared::error::{create_tuliprox_error, create_tuliprox_error_result, info_err, notify_err, str_to_io_error, to_io_error, TuliproxError, TuliproxErrorKind};
+use crate::api::model::{AppState, PlaylistStorage};
+use crate::model::normalize_release_date;
+use crate::model::{rewrite_doc_urls, PlaylistXtreamCategory, XtreamMappingOptions, XtreamSeriesEpisode};
 use crate::model::{AppConfig, ProxyUserCredentials};
 use crate::model::{Config, ConfigInput, ConfigTarget, XtreamTargetOutput};
-use crate::model::{rewrite_doc_urls, PlaylistXtreamCategory, XtreamMappingOptions, XtreamSeriesEpisode};
-use crate::model::normalize_release_date;
 use crate::repository::bplustree::{BPlusTree, BPlusTreeQuery, BPlusTreeUpdate};
 use crate::repository::indexed_document::{IndexedDocumentDirectAccess, IndexedDocumentGarbageCollector, IndexedDocumentIterator, IndexedDocumentWriter};
 use crate::repository::playlist_repository::get_target_id_mapping;
@@ -10,22 +10,23 @@ use crate::repository::storage::{get_input_storage_path, get_target_id_mapping_f
 use crate::repository::storage_const;
 use crate::repository::target_id_mapping::VirtualIdRecord;
 use crate::repository::xtream_playlist_iterator::XtreamPlaylistJsonIterator;
-use crate::utils::{bincode_deserialize, FileReadGuard};
 use crate::utils::file_reader;
+use crate::utils::json_write_documents_to_file;
 use crate::utils::open_readonly_file;
-use crate::utils::{json_write_documents_to_file};
+use crate::utils::{bincode_deserialize, FileReadGuard};
 use bytes::Bytes;
 use futures::{stream, Stream, StreamExt};
 use log::error;
 use serde::Serialize;
 use serde_json::{json, Map, Value};
+use shared::error::{create_tuliprox_error, create_tuliprox_error_result, info_err, notify_err, str_to_io_error, to_io_error, TuliproxError, TuliproxErrorKind};
+use shared::model::{PlaylistEntry, PlaylistGroup, PlaylistItem, PlaylistItemType, XtreamCluster, XtreamPlaylistItem};
+use shared::utils::{generate_playlist_uuid, get_u32_from_serde_value, hex_encode, json_iter_array};
 use std::collections::HashMap;
 use std::fs;
 use std::fs::File;
 use std::io::{BufReader, BufWriter, Error, ErrorKind, Read, Write};
 use std::path::{Path, PathBuf};
-use shared::model::{PlaylistEntry, PlaylistGroup, PlaylistItem, PlaylistItemType, XtreamCluster, XtreamPlaylistItem};
-use shared::utils::{generate_playlist_uuid, get_u32_from_serde_value, hex_encode, json_iter_array};
 
 macro_rules! cant_write_result {
     ($path:expr, $err:expr) => {
@@ -342,12 +343,87 @@ macro_rules! try_cluster {
     };
 }
 
-pub fn xtream_get_item_for_stream_id(
+fn xtream_get_item_for_stream_id_from_memory(
     virtual_id: u32,
-    app_config: &AppConfig,
+    app_state: &AppState,
     target: &ConfigTarget,
     xtream_cluster: Option<XtreamCluster>,
 ) -> Result<(XtreamPlaylistItem, VirtualIdRecord), Error> {
+    if let Some(playlist_storage) = &app_state.playlists {
+        let storage = playlist_storage.load();
+        if let Some(playlist) = storage.get(target.name.as_str()) {
+            return match playlist {
+                PlaylistStorage::M3uPlaylist(_) => {
+                    Err(str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}. It seems to be a m3u playlist")))
+                }
+                PlaylistStorage::XtreamPlaylist(xtream_storage) => {
+                    let mapping = xtream_storage.id_mapping.query(&virtual_id).ok_or_else(|| str_to_io_error(&format!("Could not find mapping for target {} and id {}", target.name, virtual_id)))?.clone();
+                    let result = match mapping.item_type {
+                        PlaylistItemType::SeriesInfo => {
+                            Ok(xtream_storage.series.query(&mapping.virtual_id)
+                                .ok_or_else(|| str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}")))?
+                                .clone())
+                        }
+                        PlaylistItemType::Series => {
+                            if let Some(item) = xtream_storage.series.query(&mapping.parent_virtual_id) {
+                                let mut xc_item = item.clone();
+                                xc_item.provider_id = mapping.provider_id;
+                                Ok(xc_item)
+                            } else {
+                                Ok(xtream_storage.series.query(&virtual_id)
+                                    .ok_or_else(|| str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}")))?
+                                    .clone())
+                            }
+                        }
+                        PlaylistItemType::Catchup => {
+                            let cluster = try_cluster!(xtream_cluster, mapping.item_type, virtual_id)?;
+                            let item = match cluster {
+                                XtreamCluster::Live => xtream_storage.live.query(&virtual_id),
+                                XtreamCluster::Video => xtream_storage.vod.query(&virtual_id),
+                                XtreamCluster::Series => xtream_storage.series.query(&virtual_id),
+                            };
+
+                            if let Some(pl_item) = item {
+                                let mut xc_item = pl_item.clone();
+                                xc_item.provider_id = mapping.provider_id;
+                                Ok(xc_item)
+                            } else {
+                                Err(str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}")))
+                            }
+                        }
+                        _ => {
+                            let cluster = try_cluster!(xtream_cluster, mapping.item_type, virtual_id)?;
+                            Ok((match cluster {
+                                XtreamCluster::Live => xtream_storage.live.query(&virtual_id),
+                                XtreamCluster::Video => xtream_storage.vod.query(&virtual_id),
+                                XtreamCluster::Series => xtream_storage.series.query(&virtual_id),
+                            }).ok_or_else(|| str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}")))?
+                                .clone())
+                        }
+                    };
+
+                    result.map(|xpli| (xpli, mapping))
+                }
+            };
+        }
+    }
+    Err(str_to_io_error(&format!("Failed to read xtream item for id {virtual_id}. No entry found.")))
+}
+
+pub fn xtream_get_item_for_stream_id(
+    virtual_id: u32,
+    app_state: &AppState,
+    target: &ConfigTarget,
+    xtream_cluster: Option<XtreamCluster>,
+) -> Result<(XtreamPlaylistItem, VirtualIdRecord), Error> {
+    if target.use_memory_cache {
+        return xtream_get_item_for_stream_id_from_memory(virtual_id,
+        app_state,
+        target,
+        xtream_cluster);
+    }
+
+    let app_config: &AppConfig = &app_state.app_config;
     let config = app_config.config.load();
     let target_path = get_target_storage_path(&config, target.name.as_str()).ok_or_else(|| str_to_io_error(&format!("Could not find path for target {}", &target.name)))?;
     let storage_path = xtream_get_storage_path(&config, target.name.as_str()).ok_or_else(|| str_to_io_error(&format!("Could not find path for target {} xtream output", &target.name)))?;
@@ -796,13 +872,13 @@ pub async fn xtream_update_input_vod_record_from_wal_file(
         let mut tmdb_id_bytes = [0u8; 4];
         let mut ts_bytes = [0u8; 8];
         let mut tree_record_index: BPlusTree<u32, InputVodInfoRecord> = BPlusTree::load(&record_path).unwrap_or_else(|_| BPlusTree::new());
-        
+
         loop {
             if reader.read_exact(&mut provider_id_bytes).is_err() {
                 break; // End of file
             }
             let provider_id = u32::from_le_bytes(provider_id_bytes);
-            
+
             if reader.read_exact(&mut tmdb_id_bytes).is_err() {
                 error!("Unexpected EOF after reading provider_id {provider_id} for VOD record.");
                 break;
@@ -817,17 +893,17 @@ pub async fn xtream_update_input_vod_record_from_wal_file(
 
             // Read the date string length as a 4-byte u32.
             let mut len_bytes = [0u8; 4];
-            if reader.read_exact(&mut len_bytes).is_err() { 
+            if reader.read_exact(&mut len_bytes).is_err() {
                 error!("Unexpected EOF when reading release_date length for VOD record with provider_id {provider_id}.");
-                break; 
+                break;
             }
             let len = u32::from_le_bytes(len_bytes) as usize;
 
             let release_date = if len > 0 {
                 let mut date_buffer = vec![0u8; len];
-                if reader.read_exact(&mut date_buffer).is_err() { 
+                if reader.read_exact(&mut date_buffer).is_err() {
                     error!("Unexpected EOF when reading release_date string for VOD record with provider_id {provider_id}.");
-                    break; 
+                    break;
                 }
                 String::from_utf8(date_buffer).ok()
             } else {
@@ -839,7 +915,7 @@ pub async fn xtream_update_input_vod_record_from_wal_file(
         }
 
         tree_record_index.store(&record_path).map_err(|err| notify_err!(format!("Could not store vod record info {err}")))?;
-        
+
         drop(reader);
         if let Err(err) = fs::remove_file(wal_path) {
             error!("Failed to delete record WAL file for vod {err}");
@@ -996,7 +1072,6 @@ pub(crate) async fn xtream_get_playlist_categories(config: &Config, target_name:
 }
 
 pub fn write_series_info_to_wal_file(provider_id: u32, ts: u64, content: &str, content_write: &mut BufWriter<&File>, record_writer: &mut BufWriter<&File>) -> std::io::Result<()> {
-
     let encoded_content = encode_info_content_for_wal_file(provider_id, content)?;
     let encoded_record = encode_series_info_record_for_wal_file(provider_id, ts);
     content_write.write_all(&encoded_content)?;
@@ -1006,7 +1081,6 @@ pub fn write_series_info_to_wal_file(provider_id: u32, ts: u64, content: &str, c
 
 
 pub fn write_vod_info_to_wal_file(provider_id: u32, content: &str, info_record: &InputVodInfoRecord, content_write: &mut BufWriter<&File>, record_writer: &mut BufWriter<&File>) -> std::io::Result<()> {
-
     let encoded_content = encode_info_content_for_wal_file(provider_id, content)?;
     let encoded_record = encode_vod_info_record_for_wal_file(provider_id, info_record)?;
     content_write.write_all(&encoded_content)?;